+++ src/components/game/coreState/utils/Functions.js
